move types to internal folder
